move drag element to separate module
